reorganizing so testing can be more granular
