+++ lite/valor_lite/text_generation/__init__.py
@@ -1,10 +1,5 @@
 from .annotation import Context, Query
-from .integrations import (
-    ClientWrapper,
-    MistralWrapper,
-    MockWrapper,
-    OpenAIWrapper,
-)
+from .integrations import ClientWrapper, MistralWrapper, OpenAIWrapper
 from .manager import Evaluator
 from .metric import Metric, MetricType
 
@@ -17,5 +12,4 @@ __all__ = [
     "ClientWrapper",
     "OpenAIWrapper",
     "MistralWrapper",
-    "MockWrapper",
 ]
